Fix for Mantis #3158: increase size of token_status to prevent narrowing when used with 0xDEADBEAD.

+++ code/network/multi_voice.cpp
@@ -109,7 +109,7 @@ int Multi_voice_stamps[MULTI_VOICE_MAX_STREAMS];
 #define MULTI_VOICE_TOKEN_INDEX_RELEASED			0xDEADBEAD		// the token has been released but the stream is still active
 
 typedef struct voice_stream {		
-	int token_status;															// status of the token (player index if a player has it) or one of the above defines
+	long token_status;															// status of the token (player index if a player has it) or one of the above defines
 	int token_stamp;															// timestamp for the MULTI_VOICE_TOKEN_TIMEOUT
 
 	short stream_from;														// id of the player the stream is coming from
